Add inline message editing (resolves #115)

+++ packages/client/js/Messages/Message/Message.js
@@ -1,11 +1,14 @@
 const { h, Component } = require('preact')
 const Icon = require('/Icon')
 const TimeAgo = require('./Timeago')
+const MessageEditor = require('../MessageEditor')
 
 // Note that this component supports both message groups (Array<Message>) *and*
 // lone messages (Message).
 class Message extends Component {
-  // TODO: markdown, better actions
+  // TODO: markdown
+
+  state = {editing: null}
 
   componentDidMount() {
     const messages = Array.isArray(this.props.msg) ? this.props.msg : [this.props.msg]
@@ -14,9 +17,12 @@ class Message extends Component {
       message.on('delete', () => this.forceUpdate())
       message.on('change', () => this.forceUpdate())
     }
+
+    this.context.pool.activeClientEE.on('login',  this.forceUpdate)
+    this.context.pool.activeClientEE.on('logout', this.forceUpdate)
   }
 
-  render({ msg, showActions = true }) {
+  render({ msg, showActions = true }, { editing }) {
     const messages = (Array.isArray(msg) ? msg : [msg]).filter(msg => !msg.deleted)
 
     if (messages.length === 0) return null
@@ -39,20 +45,44 @@ class Message extends Component {
           {authorFlair && <div class='MessageGroup-authorFlair'>{authorFlair}</div>}
           <TimeAgo date={dateCreated}/>
         </div>
-        {messages.map(message => <div
-          id={'msg-' + message.id}
-          class={(!anticipatedGroup && message.anticipated) ? 'Message --anticipated' : 'Message'}
-        >
-          <div class='Message-content'>{message.text}</div>
-          {showActions && !message.anticipated && <div class='Message-fillerLine'/>}
-          {showActions && !message.anticipated && <div class='Message-actions'>
-            <div class='Message-actions-action' title='Edit' onClick={() => message.edit(prompt('Edit message:'))}><Icon icon='edit'/></div>
-            <div class='Message-actions-action' title='Pin' onClick={() => message.pin()}><Icon icon='pin'/></div>
-            <div class='Message-actions-action' title='Delete' onClick={() => message.delete()}><Icon icon='trash'/></div>
-          </div>}
-        </div>)}
+        {messages.map(message => {
+          if (editing === message.id) {
+            return <div
+              id={'msg-' + message.id}
+              class='Message is-being-edited'
+            >
+              <MessageEditor
+                focus inline allowUploads={false}
+                content={message.text}
+
+                sendMessage={async text => {
+                  this.setState({editing: null})
+                  await message.edit(text)
+                }}
+                cancel={() => this.setState({editing: null})}
+              />
+            </div>
+          } else {
+            return <div
+              id={'msg-' + message.id}
+              class={(!anticipatedGroup && message.anticipated) ? 'Message --anticipated' : 'Message'}
+            >
+              <div class='Message-content'>{message.text}</div>
+              {showActions && !message.anticipated && <div class='Message-fillerLine'/>}
+              {showActions && !message.anticipated && <div class='Message-actions'>
+                <div class='Message-actions-action' title='Edit' onClick={this.edit(message)}><Icon icon='edit'/></div>
+                <div class='Message-actions-action' title='Pin' onClick={() => message.pin()}><Icon icon='pin'/></div>
+                <div class='Message-actions-action' title='Delete' onClick={() => message.delete()}><Icon icon='trash'/></div>
+              </div>}
+            </div>
+          }
+        })}
       </div>
     </div>
+  }
+
+  edit = message => () => {
+    this.setState({editing: message.id})
   }
 }
